+++ edge-hub/src/Microsoft.Azure.Devices.Edge.Hub.Core/HubCoreEventIds.cs
@@ -22,7 +22,7 @@ namespace Microsoft.Azure.Devices.Edge.Hub.Core
         public const int RetryingCloudProxy = EventIdStart + 1600;
         public const int NullDeviceConnectivityManager = EventIdStart + 1700;
         public const int EdgeHubConfigParser = EventIdStart + 1800;
-        public const int LocalEdgeHubConfig = EventIdStart + 1900;
+        public const int TwinConfigSource = EventIdStart + 1900;
         const int EventIdStart = 1000;
     }
 }
